+++ src/metatags/opportunity-data-mapper.js
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-export function createOpportunityData() {
+export function createOpportunityData(props = {}) {
   return {
     runbook: 'https://adobe.sharepoint.com/:w:/r/sites/aemsites-engineering/_layouts/15/doc2.aspx?sourcedoc=%7B27CF48AA-5492-435D-B17C-01E38332A5CA%7D&file=Experience_Success_Studio_Metatags_Runbook.docx&action=default&mobileredirect=true',
     origin: 'AUTOMATION',
@@ -26,6 +26,8 @@ export function createOpportunityData() {
       ],
     },
     tags: ['Traffic acquisition'],
-    data: null,
+    data: {
+      ...props,
+    },
   };
 }