+++ test/audits/metatags.test.js
@@ -14,10 +14,9 @@
 import { expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinon from 'sinon';
-import esmock from 'esmock';
 import sinonChai from 'sinon-chai';
-
 import { GetObjectCommand, ListObjectsV2Command } from '@aws-sdk/client-s3';
+import esmock from 'esmock';
 import {
   TITLE,
   DESCRIPTION,
@@ -209,6 +208,9 @@ describe('Meta Tags', () => {
       dataAccessStub = {
         Audit: {
           create: sinon.stub(),
+          AUDIT_TYPES: {
+            META_TAGS: 'meta-tags',
+          },
         },
         Configuration: {
           findLatest: sinon.stub(),
@@ -252,6 +254,38 @@ describe('Meta Tags', () => {
     });
 
     it('should process site tags and perform SEO checks', async () => {
+      const RUMAPIClientStub = {
+        createFrom: sinon.stub().returns({
+          query: sinon.stub().resolves([
+            {
+              url: 'http://example.com/blog/page1',
+              total: 100,
+              earned: 20,
+              owned: 70,
+              paid: 10,
+            },
+          ]),
+        }),
+      };
+      const mockGetRUMDomainkey = sinon.stub().resolves('mockedDomainKey');
+      const mockCalculateCPCValue = sinon.stub().resolves(2);
+      const metatagsOppty = {
+        getId: () => 'opportunity-id',
+        setAuditId: sinon.stub(),
+        save: sinon.stub(),
+        getSuggestions: sinon.stub().returns([]),
+        addSuggestions: sinon.stub().returns({ errorItems: [], createdItems: [1, 2, 3] }),
+        getType: () => 'meta-tags',
+      };
+      const site = {
+        getIsLive: sinon.stub().returns(true),
+        getId: sinon.stub().returns('site-id'),
+        getBaseURL: sinon.stub().returns('http://example.com'),
+      };
+      dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
+        isHandlerEnabledForSite: sinon.stub().returns(true),
+      });
       const topPages = [
         { getURL: 'http://example.com/blog/page1', getTopKeyword: sinon.stub().returns('page') },
         { getURL: 'http://example.com/blog/page2', getTopKeyword: sinon.stub().returns('Test') },
@@ -259,6 +293,9 @@ describe('Meta Tags', () => {
         { getURL: 'http://example.com/', getTopKeyword: sinon.stub().returns('Home') },
       ];
       dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
+      dataAccessStub.Opportunity = {
+        allBySiteIdAndStatus: sinon.stub().returns([metatagsOppty]),
+      };
       s3ClientStub.send
         .withArgs(sinon.match.instanceOf(ListObjectsV2Command).and(sinon.match.has('input', {
           Bucket: 'test-bucket',
@@ -342,7 +379,14 @@ describe('Meta Tags', () => {
 
       const addAuditStub = sinon.stub().resolves({ getId: () => 'audit-id' });
       dataAccessStub.Audit.create = await addAuditStub();
-      await auditMetaTagsRunner('http://example.com', context, { getId: () => 'site-id' });
+      const auditStub = await esmock('../../src/metatags/handler.js', {
+        '../../src/support/utils.js': { getRUMDomainkey: mockGetRUMDomainkey, calculateCPCValue: mockCalculateCPCValue },
+        '@adobe/spacecat-shared-rum-api-client': RUMAPIClientStub,
+        '@adobe/spacecat-shared-data-access': dataAccessStub,
+      });
+      const auditInstance = auditStub.default;
+      await auditInstance.runner('http://example.com', context, site);
+
       const result = await fetchAndProcessPageObject(s3ClientStub, 'test-bucket', 'scrapes/site-id/blog/page3/scrape.json', 'scrapes/site-id/', logStub);
       expect(logStub.error).to.have.been.calledWith('No Scraped tags found in S3 scrapes/site-id/blog/page3/scrape.json object');
       expect(result).to.be.null;
@@ -410,13 +454,42 @@ describe('Meta Tags', () => {
         },
       }));
       expect(addAuditStub.calledOnce).to.be.true;
+      expect(logStub.info.callCount).to.equal(6);
     });
 
     it('should process site tags and perform SEO checks for pages with invalid H1s', async () => {
+      const RUMAPIClientStub = {
+        createFrom: sinon.stub().returns({
+          query: sinon.stub().resolves([
+            {
+              url: 'http://example.com/blog/page1',
+              total: 100,
+              earned: 20,
+              owned: 70,
+              paid: 10,
+            },
+          ]),
+        }),
+      };
+      const mockGetRUMDomainkey = sinon.stub().resolves('mockedDomainKey');
+      const mockCalculateCPCValue = sinon.stub().resolves(2);
+      const auditStub = await esmock('../../src/metatags/handler.js', {
+        '../../src/support/utils.js': { getRUMDomainkey: mockGetRUMDomainkey, calculateCPCValue: mockCalculateCPCValue },
+        '@adobe/spacecat-shared-rum-api-client': RUMAPIClientStub,
+      });
+      const site = {
+        getIsLive: sinon.stub().returns(true),
+        getId: sinon.stub().returns('site-id'),
+        getBaseURL: sinon.stub().returns('http://example.com'),
+      };
       const topPages = [{ getURL: 'http://example.com/blog/page1', getTopKeyword: sinon.stub().returns('page') },
         { getURL: 'http://example.com/blog/page2', getTopKeyword: sinon.stub().returns('Test') },
         { getURL: 'http://example.com/', getTopKeyword: sinon.stub().returns('Test') }];
 
+      dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
+        isHandlerEnabledForSite: sinon.stub().returns(true),
+      });
       const getTopPagesForSiteStub = [
         { getUrl: () => 'http://example.com/blog/page1' },
         {
@@ -505,8 +578,10 @@ describe('Meta Tags', () => {
         });
       const addAuditStub = sinon.stub().resolves();
       dataAccessStub.Audit.create = await addAuditStub();
-      auditMetaTagsRunner('http://example.com', context, { getId: () => 'site-id' });
-      fetchAndProcessPageObject(s3ClientStub, 'test-bucket', 'scrapes/site-id/blog/page1/scrape.json', 'scrapes/site-id/', sinon.stub());
+
+      const auditInstance = auditStub.default;
+      await auditInstance.runner('http://example.com', context, site);
+      await fetchAndProcessPageObject(s3ClientStub, 'test-bucket', 'scrapes/site-id/blog/page1/scrape.json', 'scrapes/site-id/', sinon.stub());
 
       expect(addAuditStub.calledWithMatch({
         '/blog/page1': {
@@ -565,6 +640,7 @@ describe('Meta Tags', () => {
         },
       }));
       expect(addAuditStub.calledOnce).to.be.true;
+      expect(logStub.info.callCount).to.equal(7);
     });
 
     it('should handle gracefully if S3 object has no rawbody', async () => {
@@ -580,14 +656,14 @@ describe('Meta Tags', () => {
         })))
         .resolves({
           Contents: [
-            { Key: 'scrapes/site-id/blog/page1.json' },
+            { Key: 'scrapes/site-id/blog/page1/scrape.json' },
           ],
         });
 
       s3ClientStub.send
         .withArgs(sinon.match.instanceOf(GetObjectCommand).and(sinon.match.has('input', {
           Bucket: 'test-bucket',
-          Key: 'scrapes/site-id/blog/page1.json',
+          Key: 'scrapes/site-id/blog/page1/scrape.json',
         }))).returns({
           Body: {
             transformToString: () => '',
@@ -617,7 +693,7 @@ describe('Meta Tags', () => {
         })))
         .resolves({
           Contents: [
-            { Key: 'page1.json' },
+            { Key: 'page1/scrape.json' },
           ],
         });
 
@@ -633,9 +709,192 @@ describe('Meta Tags', () => {
           },
           ContentType: 'application/json',
         });
-      await auditMetaTagsRunner('http://example.com', context, { getId: () => 'site-id' });
+
+      const site = {
+        getIsLive: sinon.stub().returns(true),
+        getId: sinon.stub().returns('site-id'),
+        getBaseURL: sinon.stub().returns('http://example.com'),
+      };
+      const RUMAPIClientStub = {
+        createFrom: sinon.stub().returns({
+          query: sinon.stub().resolves([
+            {
+              url: 'http://example.com/blog/page1',
+              total: 100,
+              earned: 20,
+              owned: 70,
+              paid: 10,
+            },
+          ]),
+        }),
+      };
+      const mockGetRUMDomainkey = sinon.stub().resolves('mockedDomainKey');
+      const mockCalculateCPCValue = sinon.stub().resolves(2);
+      const auditStub = await esmock('../../src/metatags/handler.js', {
+        '../../src/support/utils.js': { getRUMDomainkey: mockGetRUMDomainkey, calculateCPCValue: mockCalculateCPCValue },
+        '@adobe/spacecat-shared-rum-api-client': RUMAPIClientStub,
+      });
+      const auditInstance = auditStub.default;
+      await auditInstance.runner('http://example.com', context, site);
       expect(logStub.error).to.have.been.calledWith('Failed to extract tags from scraped content for bucket test-bucket and prefix scrapes/site-id/');
     });
+
+    it('should calculate projected traffic for detected tags', async () => {
+      const RUMAPIClientStub = {
+        createFrom: sinon.stub().returns({
+          query: sinon.stub().resolves([
+            {
+              url: 'http://example.com/blog/page1',
+              total: 100,
+              earned: 20,
+              owned: 70,
+              paid: 10,
+            },
+          ]),
+        }),
+      };
+      const mockGetRUMDomainkey = sinon.stub().resolves('mockedDomainKey');
+      const metatagsOppty = {
+        getId: () => 'opportunity-id',
+        setAuditId: sinon.stub(),
+        save: sinon.stub(),
+        getSuggestions: sinon.stub().returns([]),
+        addSuggestions: sinon.stub().returns({ errorItems: [], createdItems: [1, 2, 3] }),
+        getType: () => 'meta-tags',
+      };
+      const site = {
+        getIsLive: sinon.stub().returns(true),
+        getId: sinon.stub().returns('site-id'),
+        getBaseURL: sinon.stub().returns('http://example.com'),
+      };
+      const topPages = [{ getURL: 'http://example.com/blog/page1', getTopKeyword: sinon.stub().returns('page') },
+        { getURL: 'http://example.com/blog/page2', getTopKeyword: sinon.stub().returns('Test') }];
+
+      dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
+        isHandlerEnabledForSite: sinon.stub().returns(true),
+      });
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
+      dataAccessStub.Opportunity = {
+        allBySiteIdAndStatus: sinon.stub().returns([metatagsOppty]),
+      };
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(ListObjectsV2Command).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Prefix: 'scrapes/site-id/',
+          MaxKeys: 1000,
+        })))
+        .resolves({
+          Contents: [
+            { Key: 'scrapes/site-id/blog/page1/scrape.json' },
+            { Key: 'scrapes/site-id/blog/page2/scrape.json' },
+          ],
+        });
+
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(GetObjectCommand).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Key: 'scrapes/site-id/blog/page1/scrape.json',
+        }))).returns({
+          Body: {
+            transformToString: () => JSON.stringify({
+              scrapeResult: {
+                tags: {
+                  title: 'Test Page',
+                  description: '',
+                },
+              },
+            }),
+          },
+          ContentType: 'application/json',
+        });
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(GetObjectCommand).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Key: 'scrapes/site-id/blog/page2/scrape.json',
+        }))).returns({
+          Body: {
+            transformToString: () => JSON.stringify({
+              scrapeResult: {
+                tags: {
+                  title: 'Test Page',
+                  h1: [
+                    'This is a dummy H1 that is intentionally made to be overly lengthy from SEO perspective',
+                  ],
+                },
+              },
+            }),
+          },
+          ContentType: 'application/json',
+        });
+      const addAuditStub = sinon.stub().resolves({ getId: () => 'audit-id' });
+      dataAccessStub.Audit.create = addAuditStub;
+      const mockCalculateCPCValue = sinon.stub().resolves(2);
+      const auditStub = await esmock('../../src/metatags/handler.js', {
+        '../../src/support/utils.js': { getRUMDomainkey: mockGetRUMDomainkey, calculateCPCValue: mockCalculateCPCValue },
+        '@adobe/spacecat-shared-rum-api-client': RUMAPIClientStub,
+      });
+      const auditInstance = auditStub.default;
+      await auditInstance.runner('http://example.com', context, site);
+      expect(addAuditStub.calledWithMatch({
+        auditResult: {
+          projectedTraffic: 200,
+          projectedTrafficValue: 200,
+          detectedTags: {
+            '/blog/page1': {
+              h1: {
+                seoImpact: 'High',
+                issue: 'Missing H1',
+                issueDetails: 'H1 tag is missing',
+                seoRecommendation: 'Should be present',
+              },
+              title: {
+                tagContent: 'Test Page',
+                seoImpact: 'High',
+                issue: 'Duplicate Title',
+                issueDetails: '2 pages share same title',
+                seoRecommendation: 'Unique across pages',
+                duplicates: [
+                  '/blog/page2',
+                ],
+              },
+              description: {
+                tagContent: '',
+                seoImpact: 'High',
+                issue: 'Empty Description',
+                issueDetails: 'Description tag is empty',
+                seoRecommendation: '140-160 characters long',
+              },
+            },
+            '/blog/page2': {
+              description: {
+                seoImpact: 'High',
+                issue: 'Missing Description',
+                issueDetails: 'Description tag is missing',
+                seoRecommendation: 'Should be present',
+              },
+              title: {
+                tagContent: 'Test Page',
+                seoImpact: 'High',
+                issue: 'Duplicate Title',
+                issueDetails: '2 pages share same title',
+                seoRecommendation: 'Unique across pages',
+                duplicates: [
+                  '/blog/page1',
+                ],
+              },
+              h1: {
+                tagContent: 'This is a dummy H1 that is intentionally made to be overly lengthy from SEO perspective',
+                seoImpact: 'Moderate',
+                issue: 'H1 too long',
+                issueDetails: '17 chars over limit',
+                seoRecommendation: 'Below 70 characters',
+              },
+            },
+          },
+        },
+      }));
+    }).timeout(5000);
   });
 
   describe('opportunities handler method', () => {
@@ -656,6 +915,8 @@ describe('Meta Tags', () => {
         getSuggestions: sinon.stub().returns(testData.existingSuggestions),
         addSuggestions: sinon.stub().returns({ errorItems: [], createdItems: [1, 2, 3] }),
         getType: () => 'meta-tags',
+        setData: () => {},
+        getData: () => {},
       };
       logStub = {
         info: sinon.stub(),
@@ -682,7 +943,7 @@ describe('Meta Tags', () => {
     });
 
     it('should create new opportunity and add suggestions', async () => {
-      opportunity.getType = () => 'backlinks';
+      opportunity.getType = () => 'meta-tags';
       dataAccessStub.Opportunity.create = sinon.stub().returns(opportunity);
       await opportunityAndSuggestions(auditUrl, auditData, context);
       expect(dataAccessStub.Opportunity.create).to.be.calledWith(testData.OpportunityData);