+++ test/fixtures/meta-tags-data.js
@@ -59,6 +59,8 @@ const testData = {
           },
         },
       },
+      projectedTrafficLost: 100,
+      projectedTrafficValue: 50,
     },
   },
   expectedSuggestions: [
@@ -377,7 +379,7 @@ const testData = {
     tags: [
       'Traffic acquisition',
     ],
-    data: null,
+    data: { projectedTrafficLost: 100, projectedTrafficValue: 50 },
   },
 };